feat: add Network trait

+++ src/edge_data.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, PartialEq, Clone, Default)]
 pub struct EdgeData {
-    pub weight: f32,
+    pub weight: f64,
     pub disabled: bool,
 }
